Internal research: NIGHT-lts-5 - the server long-endurance scale re-shape: wrap-coherent deltas, the u128 session ledger, and the SI ladder to quettabyte. The owner's ask ("zelynic now support for server usage... optimize, long endurance, mitigate explode data, harden and robust for future when reach limit of zelynic like possible 1 zettabyte ZB even quettabyte QB. need documented"), answered as an audit that found ONE real long-endurance bug and ONE honest ceiling, then fixed both on exactly the surfaces that carry them. The wrap-coherence bug (the hunt's find): the kernel side's booking is fetch_add - BPF atomics WRAP at u64::MAX, 18.4 EB (~4.7 years of 1-Tbps traffic through one cgroup: a REAL server horizon), documented in ebpf/src/stats.rs since improve-29 - but the userspace deltas were saturating_sub: the moment a kernel counter wrapped, every subsequent poll clamped to a ZERO delta and the cgroup went SILENT (rates zero, totals frozen) for another full 18.4 EB. The fix: wrap_coherent_delta (src/ebpf/loader.rs), modulo-2^64 subtraction - the exact inverse of the kernel's wrap, exact across a wrap whenever the true per-interval delta stays under 2^63 bytes (~9.2 EB per poll interval; a 1-Tbps link needs 2.3 YEARS to produce that - nine orders of headroom), and prev > cur cannot mean anything else (session-scoped maps, kernfs ids never reused, prev-stats only move on success). Pinned in test/ebpf/loader_wrap_tests.rs: the straddling wrap reads its true 1500 bytes, the post-wrap steady state stays alive, the coherence bound holds. The u128 session ledger (the ZB/QB widening): SessionAcc's byte legs and the footer census (grand/dl/ul) widen to u128 - the session surface sums the kernel counters' wrap-coherent deltas into totals whose honest ceiling is now ~340 million QUETTABYTES (~8.7e21 years of 1-Tbps traffic, past the SI prefix list's own end). The packet leg stays u64 (2^64 packets = ~389,000 years at 1.5 Mpps - the byte legs were the only reachable ceiling, documented). The eagle TOTAL column and the footer census render through format_bytes_wide (src/ebpf/limiter/format.rs), the u128 twin of the SI ladder: B, KB, MB, GB, TB, PB, EB, ZB, YB, RB, QB - the full 2019 prefix list, past 999.9 QB rendering the exact count (u128::MAX is "340282366.9 QB"). The u64 ladder KEEPS its 7-tier honesty for every kernel-map surface (the limiter's policy figures, the per-socket lifetime legs) - one integer width, one ladder, one truth: a formatter never renders a tier its integer cannot reach. The AVG speed pair divides through rate_bps_wide (render.rs), the u128 twin of rate_bps. The wide formatter's exactness discipline: remainder-based integer tenths (u128 has no wider register for the u64 twin's bytes*10 widening - and the carry path the first cut missed was caught by its own pins: a 10 tenth carries into the whole, the same 999.95 edge the walk enforces). The u128 widening also exposed a masked double-absorb: the footer safety pin absorbed its summary explicitly AND through the render call - masked for its whole life by the saturating add clamping the second u64::MAX onto the first; the exact fold made the doubling visible (an AVG twice the ceiling), and the pin now folds once, the way production does. The data-explosion audit re-walked at server scale: every long-lived structure re-verified fenced - the maps capped (4096 x 2), the leaderboard mirror-capped, the connections cache TTL-cleared in place, both /proc caches rebuilt behind their TTLs, the session maps recreated per run; no new growth surface, nothing to fix. Frame A/B (432c2fa vs this tree, three-run medians, 10s formal runs): bytes/frame byte-identical (1,919.0 - the wide ladder's lower tiers ARE the u64 ladder, pinned), fps -2.4 percent and every visual metric sub-noise inside the container's +/-7 percent band (the single first-run -11.6 percent was build warm-up, the exact reason the protocol requires medians) - recorded in PERFORMANCE.md. 388 + 34 tests green with the ebpf feature (9 new pins: 4 wrap-coherence, 4 wide-ladder, and the ZB census render - the footer census pin renders a 233-day 1-Tbps session as "total usage internet in 232d:18h = 1.0 ZB"); USAGE carries the scale-contract paragraph (the kernel counters' wrap, the wrap-coherent deltas, the u128 ledger, the two ladders), STABILITY's honest limit 5 and the killer inventory's arithmetic-endurance row updated to the u128/wrap-coherent truth, and the u64 ladder's doc comment now points at its wide twin instead of claiming no eighth tier exists. build.sh check-all -q green locally (fmt + clippy -D warnings + tests + policy + version; cargo-audit/cargo-deny left to CI - not installed on this host, the script's own WARN-skip contract); gate-keepers 18/18

+++ src/ebpf/render/session.rs
@@ -33,6 +33,23 @@
 //! wrap-around winner); the growth bound below keeps the
 //! leaderboard's memory honest on long-uptime monitors.
 //!
+//! NIGHT-lts-5 (the server long-endurance ask: "harden and robust
+//! for future when reach limit of zelynic like possible 1 zettabyte
+//! ZB even quettabyte QB"): the byte legs widen to u128. The
+//! kernel's per-cgroup counters are u64 by BPF-map contract and
+//! WRAP at 18.4 EB (ebpf/src/stats.rs — and the userspace deltas
+//! went wrap-coherent with them the same night, loader.rs), so a
+//! 1-Tbps-backed cgroup feeding the monitor for ~4.7 years hands
+//! the accumulator wrapping deltas forever; the u128 accumulator
+//! folds them into a session total whose honest ceiling is now
+//! u128's — 3.4e38 bytes, ~340 million QB, ~8.7e21 years of
+//! 1-Tbps traffic — past quettabyte, the SI prefix list's end.
+//! The packet leg stays u64 (2^64 packets is ~389,000 years at
+//! 1.5 Mpps line rate — the byte legs were the only reachable
+//! ceiling). The renders run through format_bytes_wide, the u128
+//! twin of the SI ladder, so the board and the footer census can
+//! SAY "1.0 ZB" — and mean it.
+//!
 //! NIGHT-boost-14: the takeover-BLINK bookkeeping is gone — the
 //! owner's eye-strain call. A takeover still re-crowns (the rank
 //! order moves), but rank 1 reads by its static champion red, never
@@ -57,24 +74,33 @@ use crate::ebpf::loader::CounterSummary;
 /// Per-cgroup traffic accumulated since the monitor started.
 #[derive(Debug, Default, Clone, Copy, PartialEq, Eq)]
 pub(crate) struct SessionAcc {
-    /// Accumulated download bytes.
-    pub dl: u64,
-    /// Accumulated upload bytes.
-    pub ul: u64,
+    /// Accumulated download bytes (u128 since NIGHT-lts-5 — the
+    /// session surface folds the kernel counters' wrap-coherent
+    /// deltas into a total that can honestly pass the exabyte;
+    /// saturating add per the boost-16 discipline, though u128
+    /// saturation needs ~8.7e21 years of 1-Tbps traffic).
+    pub dl: u128,
+    /// Accumulated upload bytes — the mirror leg, u128 for the
+    /// same reason.
+    pub ul: u128,
     /// Accumulated packets, both directions (NIGHT-engrave-4: the
     /// footer's census line counts the SESSION's packets — the same
     /// horizon as the bytes, one accumulator per frame delta, where
     /// the pre-engrave-3 census mixed a per-frame packet count with
-    /// a session cgroup count on adjacent words of one line).
+    /// a session cgroup count on adjacent words of one line). Stays
+    /// u64 (NIGHT-lts-5): 2^64 packets is ~389,000 years at 1.5 Mpps
+    /// — the byte legs were the only reachable ceiling.
     pub pkt: u64,
 }
 
 impl SessionAcc {
     /// Combined accumulated bytes (the ranking key). Saturating
-    /// (NIGHT-boost-16): both legs near u64::MAX must read as
-    /// "saturated maximum", not panic in debug or wrap in release.
+    /// (NIGHT-boost-16): the legs near the ceiling must read as
+    /// "saturated maximum", never panic in debug. The u128 widening
+    /// (NIGHT-lts-5) moves that ceiling past the quettabyte —
+    /// unreachable by ~22 orders of magnitude.
     #[must_use]
-    fn total(self) -> u64 {
+    fn total(self) -> u128 {
         self.dl.saturating_add(self.ul)
     }
 }
@@ -127,19 +153,23 @@ impl SessionState {
     /// Fold one poll's deltas into the leaderboard. An empty summary
     /// (a quiet frame, or the one-frame tolerance for a transient
     /// map-read error) folds nothing — the board holds its rows. The
-    /// fold itself is saturating (an accumulator that reaches
-    /// u64::MAX stays there — the SI formatter renders the ceiling
-    /// as 18.4 EB, saturation not a wrap), and the entry count is bounded by
-    /// [`MAX_TRACKED_CGROUPS`] (NIGHT-boost-16): a cgroup the kernel
-    /// never counted cannot rank.
+    /// fold itself is saturating (the byte legs' u128 saturation
+    /// ceiling is past the quettabyte — NIGHT-lts-5; the packet leg's
+    /// u64 ceiling is ~389,000 years of line rate), and the entry
+    /// count is bounded by [`MAX_TRACKED_CGROUPS`] (NIGHT-boost-16):
+    /// a cgroup the kernel never counted cannot rank.
     pub(crate) fn absorb(&mut self, summary: &CounterSummary) {
         for c in &summary.cgroups {
             if !self.admits(c.cgroup_id) {
                 continue;
             }
             let entry = self.acc.entry(c.cgroup_id).or_default();
-            entry.dl = entry.dl.saturating_add(c.ingress_bytes);
-            entry.ul = entry.ul.saturating_add(c.bytes);
+            // The u128 folds (NIGHT-lts-5): the deltas arrive as u64
+            // (the kernel counters' width) and widen at the fold —
+            // wrap-coherent deltas summed into a total that can
+            // pass the exabyte the kernel counters cannot.
+            entry.dl = entry.dl.saturating_add(u128::from(c.ingress_bytes));
+            entry.ul = entry.ul.saturating_add(u128::from(c.bytes));
             // Both directions' packets fold into one session counter
             // (NIGHT-engrave-4) — saturating like every accumulation
             // surface in this path.
